requirements.txt
# OrdinaryEpic — Python backend dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Google AI
google-genai>=0.8.0

# Google Cloud / Vertex AI (for Veo 3.1)
google-cloud-aiplatform>=1.70.0
vertexai>=1.70.0

# HTTP client (for Notion API)
httpx>=0.27.0

# Env / config
python-dotenv>=1.0.1
